+++ src/modules/keyboardmanager/KeyboardManagerEditor/EditKeyboardWindow.cpp
@@ -14,6 +14,7 @@
 #include <dll/Generated Files/resource.h>
 
 #include "EditKeyboardWindow.h"
+#include "ErrorTypes.h"
 #include "SingleKeyRemapControl.h"
 #include "KeyDropDownControl.h"
 #include "XamlBridge.h"
